Fix remaining type mismatches in payments and adjustments
- booking_payment_screen: use double for remaining, payments confirmations, extension amounts; adjust helper to accept num
- booking_checkout_screen already uses double amounts
- booking_price_adjustments_adapter: parse amount as double
- booking_price_adjustment_service: store adjustment amount as double
- restore_fix_service: clamp discounted rate as double

+++ mobile/lib/services/booking_price_adjustment_service.dart
@@ -220,7 +220,7 @@ class BookingPriceAdjustmentService {
       bookingLocalUuid: Value(bookingLocalUuid),
       bookingLocalId: Value(booking.id),
       adjustmentType: Value(type.value),
-      amount: Value(amount),
+      amount: Value(amount.toDouble()),
       effectiveHotelDay: Value(effectiveHotelDay),
       endHotelDay: Value(endHotelDay),
       isActive: const Value(true),
